refactor: fix clippy warnings across workspace
- Remove unnecessary cast, redundant closures, manual let bindings
- Derive Default instead of manual impls (config, backend, detector, plugin)
- Use strip_prefix, unwrap_or, is_some_and, is_none_or idioms
- Change &PathBuf to &Path in function signatures
- Use while let, enumerate, next_back, collapsed if statements
- Add type alias for complex type in rpc_client
- Mark unused-but-reserved methods with #[allow(dead_code)]
- Fix doc comment indentation

+++ crates/omnish-pty/src/proxy.rs
@@ -64,7 +64,7 @@ impl PtyProxy {
     }
 
     pub fn child_pid(&self) -> i32 {
-        self.child_pid.as_raw() as i32
+        self.child_pid.as_raw()
     }
 
     pub fn read(&self, buf: &mut [u8]) -> Result<usize> {
